+++ apps/web-giddh/src/app/invoice/proforma/proforma-list.component.html
@@ -721,12 +721,12 @@
         >
             <div class="modal-dialog modal-md">
                 <div class="modal-content">
-                    <delete-role-confirmation-model
+                    <delete-voucher-confirmation-modal
                         [module]="voucherType + 'list'"
                         (confirmDeleteEvent)="deleteVoucher()"
                         (closeModelEvent)="toggleConfirmationModel()"
                     >
-                    </delete-role-confirmation-model>
+                    </delete-voucher-confirmation-modal>
                 </div>
             </div>
         </div>
