+++ apps/web-giddh/src/app/proforma-invoice/proforma-invoice.module.ts
@@ -32,6 +32,7 @@ import { AdvanceReceiptAdjustmentModule } from '../shared/advance-receipt-adjust
 import { HasFocusDirectiveModule } from '../shared/helpers/directives/has-focus/has-focus.module';
 import { ReplacePipeModule } from '../shared/helpers/pipes/replace/replace.module';
 import { NgxBootstrapSwitchModule } from 'ngx-bootstrap-switch';
+import { VoucherAddBulkItemsModule } from '../shared/voucher-add-bulk-items/voucher-add-bulk-items.module';
 
 @NgModule({
     imports: [
@@ -64,7 +65,8 @@ import { NgxBootstrapSwitchModule } from 'ngx-bootstrap-switch';
         AdvanceReceiptAdjustmentModule,
         HasFocusDirectiveModule,
         ReplacePipeModule,
-        NgxBootstrapSwitchModule.forRoot()
+        NgxBootstrapSwitchModule.forRoot(),
+        VoucherAddBulkItemsModule
     ],
     exports: [ProformaInvoiceComponent, SalesAddStockComponent, AsideMenuProductServiceComponent],
     declarations: [
