+++ apps/web-giddh/src/app/purchase/purchase-order-preview/purchase-order-preview.component.html
@@ -263,12 +263,12 @@
 >
     <div class="modal-dialog modal-md">
         <div class="modal-content">
-            <delete-role-confirmation-model
+            <delete-voucher-confirmation-modal
                 [module]="'purchaseorder'"
                 (confirmDeleteEvent)="deleteItem()"
                 (closeModelEvent)="closeConfirmationPopup()"
             >
-            </delete-role-confirmation-model>
+            </delete-voucher-confirmation-modal>
         </div>
     </div>
 </div>
