+++ apps/web-giddh/src/app/shared/send-email-invoice/send-email-invoice.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { SendEmailInvoiceComponent } from './send-email-invoice.component';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { SharedModule } from '../shared.module';
+import { TranslateDirectiveModule } from '../../theme/translate/translate.directive.module';
 
 @NgModule({
     imports: [
         FormsModule,
         CommonModule,
-        SharedModule
+        TranslateDirectiveModule
     ],
     exports: [
         SendEmailInvoiceComponent
